split out serialization to another file
